Tweak the dataset and resource edit pages
Mostly replacing the words dataset and resource with package and file.

A lot of this could be done in translations, instead of using form
customization.

Also hide the number of followers and the follow button from the sidebar
on dataset pages, and hide the file format from the sidebar when editing
resources.

There's still a javascript module with the words "Reorder resources" in
it, this has been changed in CKAN core (custom branch) for now but could
be done in a translation.