Compute temporal extents for date and time columns
If the type of a field is set to date, time or datetime when calling
resource_schema_field_create() or _update() then a temporal extent is computed
and added to the field.

The temporal extent of a column is added to its corresponding field in the
JSON schema with key "temporal_extent", with the value in ISO 8601 time
interval format.

* Add ResourceFileDoesNotExistException, this is now raised by
  get_path_to_resource_file() if the resource doesn't have a file in the
  FileStore (previously it returned None)

* infer_schema_from_csv_file() now transforms numpy booleans into standard
  Python ones (I was getting a crash because numpy booleans aren't JSON
  serializable)

* Some refactoring for code sharing