Implement new metadata viewer
Remove the SimpleCSVPreview plugin, we're no longer using CKAN's preview
infrastructure because it uses an iframe, replace it with our own snippet
instead.

Add a new template snippet that renders the CSV table and the selected column's
metadata side-by-side, and lets you click on columns to change which column's
metadata is shown.

This is done without JS currently, so there's a different URL for showing each
column's metadata, URLs like /package/id/file/id/schema/index are used for
this.

The old View Schema pages are gone. The old Edit Schema pages are still there,
but they're now accessed by an Edit Schema button where the View Schema button
used to be, rather than by clicking on the columns in the CSV preview. This
schema editing should be replaced as well.

Tests for this still need to be written, may take some ideas from the old
SimpleCSVPreview tests.