Implement changing between columns in the resource editor
If the user clicks on a column in the CSV preview on the resource edit page,
the page reloads with that column selected and any values entered into form
fields on previously selected columns are maintained.

Lots of logic added to the custom package controller to handle this.

schema_fields is now passed from the controller to the datapackager_csv.html
snippet, and used to fill the values of the form fields. schema_fields is
either the resource schema values from the database, or those previously
submitted by the user during this editing "session".

Renamed selected_row in the templates and controller to selected_column.

Update some tests.