Implement validation error handling in the resource schema editor
Every time the resource edit form is submitted - regardless of whether
submitted by clicking Save or just by clicking to go to another column of the
CSV - the resource schema field data entered into the form is validated. Any
errors found are sent to the template as schema_errors, and rendered next to
their form fields.

The first column with errors is always shown as the selected column, unless
there are no columns with errors, then we show the column that the user
clicked on. So as long as there are validation errors with the current column,
the user is not allowed to move to another column.

*Lots* of complicated new logic in the controller.

Some refactoring of the controller to move things into small helper functions
and remove duplication.

Some CSS tweaks to make errors look good on the form.

Added validate_only param to resource_schema_field_update() so that it can
be used to only validate the data. _create() will need the same param.