[#26] Export the dataset's extras
Currently we put the extras inside the root-level of the datapackage.json.
That's good, because we set the pattern of putting in extras whatever we don't
recognize, but it also creates an issue: what if an extra's key clashes with a
datapackage key? For example, what if there's an extra with a key "sources"?

CKAN validates that there're no extras with keys that clashes with its values
(e.g. name, author, etc.), but there're many fields in datapackage that have
different names (or don't exist) in CKAN (e.g. sources and license), so CKAN
doesn't prohibits them.

We could either add those validations ourselves or not add the extras into the
datapackage.json's root level.