[#39] Fix bug where uploaded datapackage wasn't being used
This problem was introduced in 5b5f4e8d. There, I've changed the test if a file
was uploaded from testing if `upload is None` to testing if `upload` is truthy.
It turns out `cgi.FieldStorage` is falsy, so this test was failing, and we were
instead using the `url` attribute.

With this patch, I test that `upload.file` and `upload.file.read` exists.